Move game to local
