+++ internal/game/game.go
@@ -1,33 +1,15 @@
 package game
 
-import (
-	"context"
-	"io"
+import "errors"
+
+type GameServer interface {
+	Start() error
+	IsRunning() (bool, error)
+	Stop() error
+	WriteFile(path string, data []byte) error
+	ReadFile(path string) ([]byte, error)
+}
+
+var (
+	ErrIsDir = errors.New("path is a directory")
 )
-
-const (
-	pidFile     = "prbf2.pid"
-	updaterPath = "mods/pr/bin"
-)
-
-type Server struct {
-	Path       string
-	processPID *int
-}
-
-func Open(path string) (*Server, error) {
-	s := &Server{
-		Path: path,
-	}
-
-	err := s.retrieveProcessPID()
-	if err != nil {
-		return nil, err
-	}
-
-	return s, nil
-}
-
-func (s *Server) Update(ctx context.Context, outW io.Writer, inR io.Reader, errW io.Writer) error {
-	return s.update(ctx, outW, inR, errW)
-}
