+++ internal/game/local/game_test.go
@@ -1,4 +1,4 @@
-package game
+package local
 
 import (
 	"os"
