+++ internal/game/local/process_windows.go
@@ -1,6 +1,6 @@
 //go:build windows
 
-package game
+package local
 
 import (
 	"os"
