+++ internal/game/local/update_windows.go
@@ -1,6 +1,6 @@
 //go:build windows
 
-package game
+package local
 
 import (
 	"context"
